Creates final release tarball, permits naming it
Our clamav-release Github release process depends on the Bosh Final
Release creation process also creating a release tarball. This commit
unconditionally creates that tarball in the 'release_metadata'
directory, and allows the operator to specify a "base" name for that
tarball.

So, for example, a "base" name of 'basename' and a release version of
6.5.4 would see a tarball created in
'release_metadata/basename-6.5.4.tgz'. Operators who are using this
script can choose to ignore this release tarball.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>